Update app.component.html
Previous build was no successful due to the issue in the code line. For the testing of deleting the tasks added, we had mentioned one selector to select the button. But had missed to add that class in the html document. Due to which the build was failed. Correcting it by maintaining the same.